feat:【PAC模板】流水线模板支持PAC特性 #11414 模版状态转换

+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/api/template/v2/UserPipelineTemplateV2Resource.kt
@@ -79,7 +79,7 @@ interface UserPipelineTemplateV2Resource {
         templateId: String,
         @Parameter(description = "请求体", required = true)
         request: PipelineTemplateDraftSaveReq
-    ): Result<Long>
+    ): Result<Int>
 
     @Operation(summary = "获取模板列表")
     @POST
