+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/api/template/v2/UserPipelineTemplateV2ResourceImpl.kt
@@ -86,7 +86,7 @@ class UserPipelineTemplateV2ResourceImpl(
         projectId: String,
         templateId: String,
         request: PipelineTemplateDraftSaveReq
-    ): Result<Long> {
+    ): Result<Int> {
         logger.info("save template draft {}|{}|{}|{}", userId, projectId, templateId, request)
         permissionService.checkPipelineTemplatePermissionWithMessage(
             userId = userId,
